product slider image add api change

+++ app/Http/Controllers/Admin/RolesController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Auth;
 use App\Models\Auth\Admin;
 use Illuminate\Http\Request;
 use Spatie\Permission\Models\Role;
@@ -18,6 +19,10 @@ class RolesController extends Controller
      */
     public function index()
     {
+        if(is_null(Auth::guard('admin')->user()) || !Auth::guard('admin')->user()->can('role.view')){
+            abort(403, 'You have no access this page.');
+        };
+
         $roles = Role::get();
         return view('admin.roles.index', compact('roles'));
     }
@@ -29,6 +34,9 @@ class RolesController extends Controller
      */
     public function create()
     {
+        if(is_null(Auth::guard('admin')->user()) || !Auth::guard('admin')->user()->can('role.create')){
+            abort(403, 'You have no access this page.');
+        };
         $permissions = Permission::all();
         $permission_groups = Admin::getpermissionGroups();
 
@@ -43,6 +51,9 @@ class RolesController extends Controller
      */
     public function store(Request $request)
     {
+        if(is_null(Auth::guard('admin')->user()) || !Auth::guard('admin')->user()->can('role.create')){
+            abort(403, 'You have no access this page.');
+        };
         $request->validate([
             'name' => 'required |unique:roles'
         ],[
@@ -70,6 +81,9 @@ class RolesController extends Controller
      */
     public function show($id)
     {
+        if(is_null(Auth::guard('admin')->user()) || !Auth::guard('admin')->user()->can('role.view')){
+            abort(403, 'You have no access this page.');
+        };
         return "ok";
     }
 
@@ -81,6 +95,9 @@ class RolesController extends Controller
      */
     public function edit($id)
     {
+        if(is_null(Auth::guard('admin')->user()) || !Auth::guard('admin')->user()->can('role.update')){
+            abort(403, 'You have no access this page.');
+        };
         $role = Role::findById($id, 'admin');
         $permissions = Permission::all();
         $permission_groups = Admin::getpermissionGroups();
@@ -97,6 +114,9 @@ class RolesController extends Controller
      */
     public function update(Request $request, $id)
     {
+        if(is_null(Auth::guard('admin')->user()) || !Auth::guard('admin')->user()->can('role.update')){
+            abort(403, 'You have no access this page.');
+        };
         $request->validate([
             'name' => 'required'
         ],[
@@ -125,6 +145,9 @@ class RolesController extends Controller
      */
     public function destroy($id)
     {
+        if(is_null(Auth::guard('admin')->user()) || !Auth::guard('admin')->user()->can('role.delete')){
+            abort(403, 'You have no access this page.');
+        };
         $role = Role::findById($id,'admin');
         if (!is_null($role)) {
             $role->delete();
